requirements.txt
APScheduler==3.10.4
async-timeout==4.0.3
blinker==1.7.0
cachelib==0.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-APScheduler==1.13.1
Flask-Caching==2.1.0
Flask-Session==0.6.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.2
requests==2.31.0
schedule==1.2.1
six==1.16.0
spotipy==2.23.0
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.1
